+++ nlpmodels/utils/__init__.py
@@ -1,5 +1,9 @@
 # pylint: disable=C0111
 
-__all__ = ["train", "skipgram_dataset", "tokenizer",
-           "vocabulary", "utils", "dataset", "label_smoother",
-           "optims", "transformer_dataset", "transformer_batch", "gpt_sampler"]
+__all__ = ["train",
+           "tokenizer",
+           "vocabulary",
+           "utils",
+           "label_smoother",
+           "optims",
+           "gpt_sampler"]
