+++ nlpmodels/utils/elt/gpt_dataset.py
@@ -8,7 +8,7 @@ import torch
 from torch.utils.data import DataLoader
 from torchtext.experimental.datasets import WikiText2
 
-from nlpmodels.utils.dataset import AbstractNLPDataset
+from nlpmodels.utils.elt.dataset import AbstractNLPDataset
 from nlpmodels.utils.vocabulary import NLPVocabulary
