+++ nlpmodels/utils/elt/text_cnn_dataset.py
@@ -2,14 +2,13 @@
 This module contains the cnn-text Dataset and Dataloader.
 """
 from typing import Tuple, Any, List
-import numpy as np
 
 import torch
 from torch.utils.data import DataLoader
 
 from torchtext.experimental.datasets import IMDB
 
-from nlpmodels.utils.dataset import AbstractNLPDataset
+from nlpmodels.utils.elt.dataset import AbstractNLPDataset
 from nlpmodels.utils.tokenizer import tokenize_corpus_basic
 from nlpmodels.utils.vocabulary import NLPVocabulary
