+++ tests/test_transformer_dataset.py
@@ -2,7 +2,7 @@ from argparse import Namespace
 
 import torch
 
-from nlpmodels.utils.transformer_dataset import TransformerDataset
+from nlpmodels.utils.elt.transformer_dataset import TransformerDataset
 from nlpmodels.utils.vocabulary import NLPVocabulary
